feat(utils,-email_manager): Emailing on crucial events
BREAKING CHANGE: The client requested if there was a way to send emails upon the banning / unbanning of collaborators as well as the publishing/unpublishing of documents. To comply, an email manager class was created to send emails when provided with a valid google account. The client is able to enable or disable the sending of emails by changing the environment variable ENABLE_EMAIL. Also, the client will need to provide a valid gmail account to use the email_manager.